Adição de proteção contra alterações indevidas no ID de agente e no método apresentado

+++ Multas/Multas/Controllers/AgentesController.cs
@@ -59,6 +59,11 @@ namespace Multas.Controllers
         }
 
         // GET: Agentes/Edit/5
+        /// <summary>
+        /// mostra na view os dados de um agente para posterior, eventual, remoção.
+        /// </summary>
+        /// <param name="id">Identificador do agente a remover</param>
+        /// <returns></returns>
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -92,24 +97,73 @@ namespace Multas.Controllers
         // GET: Agentes/Delete/5
         public ActionResult Delete(int? id)
         {
+            // o id do agente não foi fornecido
+            // não é possivel procurar o Agente
+            // o que devo fazer?
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                // opção por defeito do template:
+                // return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+
+                /// e não há ID de Agente, uma das duas coisas aconteceu:
+                ///     - há um erro nos links da aplicação
+                ///     - há um 'chico experto' a fazer asneiras no URL
+                return RedirectToAction("Index");
             }
-            Agentes agentes = db.Agentes.Find(id);
-            if (agentes == null)
+            // Procura os dados do agente cujo ID é fornecido.
+            Agentes agente = db.Agentes.Find(id);
+
+            // se o agente não for encontrado
+            /// ou há um erro,
+            /// ou há um 'chico experto' a fazer asneiras no URL
+            if (agente == null)
             {
-                return HttpNotFound();
+                return RedirectToAction("Index");
             }
-            return View(agentes);
+
+            // Para o caso do utilizador alterar o valor da action no post, de forma fraudulenta.
+            // Para isso vou guardar o valor do ID do agente.
+            // - Guardar o ID do Agente num cookie cifrado.
+            // - Guardar o ID numa variavél de sessão. (quem estiver a usar o Asp.Net CORE já não tem esta ferramenta.)
+            Session["idAgente"] = agente.ID;
+            Session["Metodo"] = "Agentes/Delete";
+
+            // Envia para a view os dados do agente em conta
+            return View(agente);
         }
 
         // POST: Agentes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int? id)
         {
+            if (id == null)
+            {
+                // se entrei aqui, é porque há um erro
+                // não se sabe o ID do agente a remover
+                return RedirectToAction("Index");
+            }
+            // Avaliar se o ID do agente que é fornecido é o mesmo ID do agente que foi apresentado no ecrã
+            if (id != (int)Session["idAgente"])
+            {
+                return RedirectToAction("Index");
+            }
+            // Avaliar se o metodo que é fornecido é o mesmo método que foi apresentado no ecrã
+            string operacao = "Agentes/Delete";
+            if (operacao != (string)Session["Metodo"])
+            {
+                return RedirectToAction("Index");
+            }
+
             Agentes agente = db.Agentes.Find(id);
+
+
+
+            if(agente == null)
+            {
+                // não foi possivel encontrar o Agente
+                RedirectToAction("Index");
+            }
             try
             {
                 db.Agentes.Remove(agente);
